#170: Fix for Elasticsearch index creation and disabling Elasticsearch

+++ c2mon-server/c2mon-server-elasticsearch/src/main/java/cern/c2mon/server/elasticsearch/tag/config/TagConfigDocumentIndexer.java
@@ -83,6 +83,10 @@ public class TagConfigDocumentIndexer {
   }
 
   public void removeTagConfig(TagConfigDocument tag) {
+    if (!Indices.exists(this.configIndex)) {
+      return;
+    }
+
     DeleteRequest deleteRequest = new DeleteRequest(configIndex, TYPE,
             String.valueOf(tag.getId())).routing(tag.getId());
